requirements.txt
-i https://pypi.org/simple
certifi==2024.6.2 ; python_version >= '3.6'
charset-normalizer==3.3.2 ; python_full_version >= '3.7.0'
idna==3.7 ; python_version >= '3.5'
python-dateutil==2.9.0.post0
requests==2.32.3 ; python_version >= '3.8'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
so4t-api==0.1.5
urllib3==2.2.2 ; python_version >= '3.8'
